requirements.txt
praw
pushshift.py
requests
beautifulsoup4
sentence-transformers
hdbscan
transformers
torch
moviepy
openai-whisper
ultralytics
pandas
geopandas
plotly
vaderSentiment
